Make the pkcs12 export option include the chain
Currently, when using the pkcs12 method on a Certificate object, it does
not export the chain in the pkcs12 bundle which can lead to serving the
leaf certificate without the intermediate if you load the pkcs12
directly into a server. This change causes the pkcs12 method to properly
export the chain with the cert and key.

I've also included a fix for a second bug in the Certificate by_issuance
constructor. If there ever were a case where there was more than one
intermediate, this call would have failed. This now properly splits the
chain into the two available parameters in all situations.