Support ECDSA / RSA 2048+ certificates
This patch adds three command-line options for `acmesmith order` to
specify the public key algorithm for the certificate to be requested.

- `--key-type` specifies the type of generated keypair.
  Supported values are `rsa` and `ecdsa` (default: `rsa`)
- `--rsa-key-size` specifies the RSA modulus size in bits (default: `2048`)
- `--elliptic-curve` specifies the ECDSA curve group.
  Possible values are up to the linked OpenSSL and the CA. (default: `prime256v1`)

The default values match the current behaviour (RSA 2048 bits).

When a certificate is initially ordered, Acmesmith generates a keypair
of the type and the key size / elliptic curve given in the command line.
On subsequent renewals, a fresh keypair of the same algorithm and the
size as the existing public key is used to order a new certificate.